Update Mintbase component imports
