+++ apps/BosGenaDrop/widget/Mintbase/App/Profile/Index.jsx
@@ -1,10 +1,14 @@
 const accountId = props.accountId ?? context.accountId;
 
 const isConnected = context.accountId === accountId;
-const { MbModal, MbInputField } = VM.require(
+const { MbModal } = VM.require(
   "bos.genadrop.near/widget/Mintbase.components"
 ) || {
   MbModal: () => <></>,
+};
+const { MbInputField } = VM.require(
+  "bos.genadrop.near/widget/Mintbase.MbInput"
+) || {
   MbInputField: () => <></>,
 };
