+++ apps/BosGenaDrop/widget/Mintbase/MbInput.jsx
@@ -1,7 +1,3 @@
-const { getInputLabelFontType, getFontType } = VM.require(
-  "bos.genadrop.near/widget/Mintbase.Theme"
-);
-
 const { MbCharCounter } = VM.require(
   "bos.genadrop.near/widget/Mintbase.components"
 );
@@ -12,46 +8,23 @@ const EControlStatus = {
   INVALID: "invalid",
 };
 
-const {
-  id,
-  required,
-  disabled,
-  customStyle,
-  placeholder,
-  label, //string
-  hasPercentageLabel,
-  value,
-  type,
-  hasIcon, //boolean
-  maxChars, // number
-  defaultValue, // number | undefined
-  onChange,
-  customIcon, // JSX.Element
-  ...props
-} = props;
-
 const inputSize = props.inputSize || "medium";
 const controlStatus = props.controlStatus || "normal";
-const initialCounter = props.initialCounter || 0;
-const mode = props.mode || Storage.get("mode");
-
-const IsDarkModeOn = mode === "dark";
-
-const [count, setCount] = useState(initialCounter);
-
-const getIconSize = () => {
-  return inputSize === "big" ? "24px" : "20px";
-};
 
 const Label = styled.div`
   display: block;
   margin-bottom: 8px;
-  ${IsDarkModeOn ? "color: white;" : ""}
-  ${getInputLabelFontType(inputSize)}
+  color: #000;
+  &.dark {
+    color: #fff;
+  }
 `;
 
 const Asterisk = styled.span`
-  color: ${IsDarkModeOn ? "#ED5A5A" : "#C74C4C"};
+  color: #c74c4c;
+  &.dark {
+    color: #ed5a5a;
+  }
 `;
 
 const Container = styled.div`
@@ -60,73 +33,46 @@ const Container = styled.div`
     justify-content: space-between;
     align-items: center;
     border-radius: 0.25rem;
-    ${customStyle}
     &.disabled {
-      background: ${IsDarkModeOn ? "#404252" : "#D2D4DA"};
+      background: #d2d4da;
+    }
+    &.disabled-dark {
+      background: #404252;
     }
     &.default {
-      /* @apply   focus-within:ring-1 transition-all duration-500; */
-      transition-property: all;
-      transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
-      transition-duration: 500ms;
-      background: ${IsDarkModeOn ? "#101223" : "#F3F4F8"};
+      transition: all 500ms cubic-bezier(0.4, 0, 0.2, 1);
+      background: #f3f4f8;
       :hover {
-        background: ${IsDarkModeOn ? "#070C2B" : "#EBEDFB"};
+        background: #ebedfb;
       }
       :focus-within {
-        --tw-ring-offset-shadow: var(--tw-ring-inset) 0 0 0
-          var(--tw-ring-offset-width) var(--tw-ring-offset-color);
-        --tw-ring-shadow: var(--tw-ring-inset) 0 0 0
-          calc(1px + var(--tw-ring-offset-width)) var(--tw-ring-color);
-        box-shadow: var(--tw-ring-offset-shadow), var(--tw-ring-shadow),
-          var(--tw-shadow, 0 0 #0000);
+        box-shadow: 0 0 0 1px #c5d0ff;
+      }
+    }
+    &.default-dark {
+      transition: all 500ms cubic-bezier(0.4, 0, 0.2, 1);
+      background: #101223;
+      :hover {
+        background: #070c2b;
+      }
+      :focus-within {
+        box-shadow: 0 0 0 1px #c5d0ff;
       }
     }
     &.empty:focus-within {
-      ${IsDarkModeOn
-        ? ` --tw-ring-opacity: 1;
-            --tw-ring-color: var(--blue-100-35);`
-        : `
-            --tw-ring-opacity: 1;
-            --tw-ring-color:  var(--blue-300-35);
-        `}
+      box-shadow: 0 0 0 1px #1aa7ec;
     }
     &.valid {
-      --tw-ring-offset-shadow: var(--tw-ring-inset) 0 0 0
-        var(--tw-ring-offset-width) var(--tw-ring-offset-color);
-      --tw-ring-shadow: var(--tw-ring-inset) 0 0 0
-        calc(1px + var(--tw-ring-offset-width)) var(--tw-ring-color);
-      ${IsDarkModeOn
-        ? ` --tw-ring-opacity: 1;
-            --tw-ring-color: var(--success-100);`
-        : `
-            --tw-ring-opacity: 1;
-            --tw-ring-color: var(--success-300);
-        `}
-      box-shadow: var(--tw-ring-offset-shadow), var(--tw-ring-shadow),
-        var(--tw-shadow, 0 0 #0000);
+      box-shadow: 0 0 0 1px #1aa7ec;
     }
     &.invalid {
-      ${IsDarkModeOn
-        ? ` --tw-ring-opacity: 1;
-            --tw-ring-color: var(--error-100);`
-        : `
-            --tw-ring-opacity: 1;
-            --tw-ring-color: var(--error-300);
-        `}
-      box-shadow: var(--tw-ring-offset-shadow), var(--tw-ring-shadow),
-        var(--tw-shadow, 0 0 #0000);
+      box-shadow: 0 0 0 1px #ed5a5a;
+      :focus-within {
+        box-shadow: 0 0 0 1px #ed5a5a;
+      }
     }
   }
-  .input-wrapper.big {
-    padding: 12px;
-    max-height: 40px;
-    @media (min-width: 480px) {
-      padding: 16px;
-      max-height: 48px;
-    }
-  }
-  .input-wrapper.medium {
+  .input-wrapper {
     padding: 10px;
     max-height: 34px;
     @media (min-width: 480px) {
@@ -134,149 +80,155 @@ const Container = styled.div`
       max-height: 40px;
     }
   }
-  .input-wrapper.small {
-    padding: 10px;
-    max-height: 34px;
-  }
-  .input-wrapper {
-    width: 100%;
-  }
 `;
 
 const InputField = styled.div`
   display: flex;
   width: 100%;
+  background: transparent;
   .input-field {
     outline: none;
     border: none;
     background: transparent;
-    color: ${isDarkModeOn ? "#fff" : "black"};
     width: 100%;
-    ${getFontType(inputSize)}
     :focus {
-      outline: 2px solid transparent;
+      outline: none;
       outline-offset: 2px;
     }
+  }
+  .dark-btn {
+    color: #fff;
     ::placeholder {
-      color: ${isDarkModeOn ? "white" : "#777986"};
+      color: var(--gray-400, #9496a1);
+    }
+  }
+  .light-btn {
+    color: #000;
+    ::placeholder {
+      color: #777986;
     }
   }
   .percentage-label {
-    ${getFontType(inputSize)}
-    color: "#777986";
+    color: #777986;
   }
 `;
+const MbInputField = ({
+  id,
+  required,
+  disabled,
+  placeholder,
+  label, //string
+  hasPercentageLabel,
+  value,
+  type,
+  error,
+  count,
+  hasIcon, //boolean
+  maxChars, // number
+  defaultValue, // number | undefined
+  onChange,
+  customIcon, // JSX.Element
+  className,
+  style,
+  isDarkModeOn,
+}) => {
+  console.log({ disabled, isDarkModeOn, required, type, count });
 
-const handleChange = (e) => {
-  if (maxChars) {
-    setCount(e.target.value.length);
-  }
-  if (!onChange) return;
-  console.log(e);
+  const wrapperClasses =
+    disabled && isDarkModeOn
+      ? "disabled-dark"
+      : disabled && !isDarkModeOn
+      ? "disabled"
+      : !disabled && error
+      ? "invalid"
+      : !disabled && isDarkModeOn
+      ? "default-dark"
+      : "default";
+  return (
+    <Container className={className} style={style}>
+      {label && (
+        <Label className={isDarkModeOn && "dark"} style={style}>
+          {label}
+          {required && (
+            <Asterisk className={isDarkModeOn && "dark"}> *</Asterisk>
+          )}
+        </Label>
+      )}
+      <div
+        className={`main-input input-wrapper ${wrapperClasses} ${controlStatus}`}
+      >
+        <InputField key={`input-container-${id}`}>
+          <input
+            id={id}
+            key={`input-field-${id}`}
+            required={required}
+            disabled={disabled}
+            placeholder={placeholder}
+            value={value}
+            type={type}
+            maxLength={maxChars}
+            className={`${className} ${
+              isDarkModeOn ? "dark-btn" : "light-btn"
+            }`}
+            onWheel={(e) => {
+              if (type !== "number") return;
+              e.currentTarget.blur();
+            }}
+            onChange={onChange}
+          />
+          {hasPercentageLabel && <span className="percentage-label">%</span>}
+        </InputField>
 
-  onChange(e);
+        {hasIcon && (
+          <div className="flex">
+            {controlStatus === EControlStatus.VALID ? (
+              <Widget
+                src="bos.genadrop.near/widget/Mintbase.MbIcon"
+                props={{
+                  name: "success",
+                  size: "20px",
+                  color: "success-300",
+                  darkColor: "success-100",
+                }}
+              />
+            ) : controlStatus === EControlStatus.INVALID ? (
+              <Widget
+                src="bos.genadrop.near/widget/Mintbase.MbIcon"
+                props={{
+                  name: "error",
+                  size: "20px",
+                  color: "error-300",
+                  darkColor: "error-100",
+                }}
+              />
+            ) : (
+              <Widget
+                src="bos.genadrop.near/widget/Mintbase.MbIcon"
+                props={{
+                  name: "info",
+                  size: "20px",
+                  color: "blue-300",
+                  darkColor: "blue-100",
+                }}
+              />
+            )}
+          </div>
+        )}
+        {!!customIcon && !hasIcon && customIcon}
+      </div>
+      {maxChars ? (
+        <MbCharCounter
+          counter={count}
+          inputSize={inputSize}
+          maxChars={maxChars}
+        />
+      ) : (
+        <></>
+      )}
+    </Container>
+  );
 };
 
-
-return (
-  <Container>
-    {label && (
-      <Label>
-        {label}
-        {required && <Asterisk> *</Asterisk>}
-      </Label>
-    )}
-    <div
-      className={`main-input input-wrapper ${inputSize} ${
-        disabled ? "disabled" : "default"
-      } ${controlStatus}`}
-    >
-      <InputField>
-        <input
-          disabled={disabled}
-          placeholder={placeholder}
-          type="text"
-          // value={value}
-          maxLength={maxChars}
-          required={required}
-          defaultValue={defaultValue}
-          className="input-field"
-          onWheel={(e) => {
-            if (type !== "number") return;
-            e.currentTarget.blur();
-          }}
-          onChange={handleChange}
-          {...props}
-        />
-        {hasPercentageLabel && <span className="percentage-label">%</span>}
-      </InputField>
-
-      {hasIcon && (
-        <div className="flex">
-          {controlStatus === EControlStatus.VALID ? (
-            <Widget
-              src="bos.genadrop.near/widget/Mintbase.MbIcon"
-              props={{
-                name: "success",
-                size: getIconSize(),
-                color: "success-300",
-                darkColor: "success-100",
-              }}
-            />
-          ) : controlStatus === EControlStatus.INVALID ? (
-            <Widget
-              src="bos.genadrop.near/widget/Mintbase.MbIcon"
-              props={{
-                name: "error",
-                size: getIconSize(),
-                color: "error-300",
-                darkColor: "error-100",
-              }}
-            />
-          ) : (
-            <Widget
-              src="bos.genadrop.near/widget/Mintbase.MbIcon"
-              props={{
-                name: "info",
-                size: getIconSize(),
-                color: "blue-300",
-                darkColor: "blue-100",
-              }}
-            />
-          )}
-        </div>
-      )}
-      {!!customIcon && !hasIcon && customIcon}
-    </div>
-    {maxChars ? (
-      <MbCharCounter
-        counter={count}
-        inputSize={inputSize}
-        maxChars={maxChars}
-      />
-    ) : (
-      <></>
-    )}
-  </Container>
-);
-
-/* <Widget
-          src="bos.genadrop.near/widget/MbInputField"
-          props={{
-            id: id,
-            disabled: disabled,
-            placeholder: placeholder,
-            type: "text",
-            value: count,
-            maxLength: maxChars,
-            required: required,
-            defaultValue: defaultValue,
-            className: "input-field",
-            onWheel: (e) => {
-              if (type != "number") return;
-              e.currentTarget.blur();
-            },
-            onChange: handleChange,
-            ...props,
-          }}
-        /> */
+return {
+  MbInputField,
+};
